updated topic-to-queue mapping
